Switch ColorGateway to Rolling Array For Stats
This was a change I made for demo'ing route changes over time. The
current behavior of tracking all stats makes it difficult to show
accurate distribution #s when using automated polling because you have
to make the previous results insignificant in the distribution.

This changes the stats to be stored in a rolling array that will
converge to the new distribution after maxColors queries.