Adds multiarch builds for k8s http2 example
This modifies the images used for the k8s http2 example to default support arm64
in addition to amd64. This should gracefully downgrade to existing behavior if
run from an environment that does not support cross-building.

This is done on this example as it's the "main" k8s setup for walkthroughs for
demoing multiarch k8s abilities. Further walkthroughs may be adapted in the future.